CPLP-1300 add header and test.sdk package
