+++ src/notification/CatenaX.NetworkServices.Notification.Service/Program.cs
@@ -18,8 +18,6 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-#region
-
 using CatenaX.NetworkServices.Framework.Cors;
 using CatenaX.NetworkServices.Framework.ErrorHandling;
 using CatenaX.NetworkServices.Framework.Swagger;
@@ -37,8 +35,6 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.FileProviders;
 using System.IdentityModel.Tokens.Jwt;
 using System.Text.Json.Serialization;
-
-#endregion
 
 var VERSION = "v2";
 var TAG = typeof(Program).Namespace;
@@ -131,4 +127,7 @@ app.MapControllers();
 
 app.Run();
 
-public partial class Program { }
+/// <summary>
+/// Needed for integration Test setup
+/// </summary>
+public partial class Program { }
