--- tests/administration/CatenaX.NetworkServices.Administration.Service.Tests/EnpointSetup/Paths.cs
@@ -1,7 +1,0 @@
-﻿namespace CatenaX.NetworkServices.Administration.Service.Tests.EnpointSetup
-{
-    public static class Paths
-    {
-        public static readonly string Registration = "/api/administration/registration";
-    }
-}
