+++ tests/notification/CatenaX.NetworkServices.Notification.Service.Tests/CatenaX.NetworkServices.Notification.Service.Tests.csproj
@@ -33,6 +33,7 @@
         <PackageReference Include="FakeItEasy" Version="7.3.1" />
         <PackageReference Include="FluentAssertions" Version="6.6.0" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.11.0" />
+        <PackageReference Include="xunit" Version="2.4.1" />
         <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
             <PrivateAssets>all</PrivateAssets>
