+++ tests/notification/CatenaX.NetworkServices.Notification.Service.Tests/IntegrationTests/NotificationControllerIntegrationTests.cs
@@ -1,19 +1,34 @@
-﻿using System;
+﻿/********************************************************************************
+ * Copyright (c) 2021,2022 BMW Group AG
+ * Copyright (c) 2021,2022 Contributors to the CatenaX (ng) GitHub Organisation.
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License, Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ********************************************************************************/
+
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
-using System.Net.Http;
-using System.Net.Http.Json;
 using System.Threading.Tasks;
 using CatenaX.NetworkServices.Notification.Service.Tests.EnpointSetup;
-using CatenaX.NetworkServices.PortalBackend.PortalEntities;
-using CatenaX.NetworkServices.PortalBackend.PortalEntities.Entities;
 using CatenaX.NetworkServices.PortalBackend.PortalEntities.Enums;
 using CatenaX.NetworkServices.Tests.Shared.Extensions;
 using CatenaX.NetworkServices.Tests.Shared.IntegrationTests;
 using CatenaX.NetworkServices.Tests.Shared.TestSeeds;
 using FluentAssertions;
-using Newtonsoft.Json;
 using Xunit;
 
 namespace CatenaX.NetworkServices.Notification.Service.Tests.IntegrationTests;
@@ -56,7 +71,7 @@ public class NotificationControllerIntegrationTests : IClassFixture<IntegrationT
     {
         // Arrange
         var client = _factory.CreateClient();
-        var endpoint = new NotificationEndpoint(client);
+        var endpoint = new NotificationEndpoints(client);
         
         // Act
         var response = await endpoint.NotificationCount(false);
