+++ tests/portalbackend/CatenaX.NetworkServices.PortalBackend.DBAccess.Tests/AppRepositoryTests.cs
@@ -25,6 +25,7 @@ using CatenaX.NetworkServices.PortalBackend.DBAccess.Repositories;
 using CatenaX.NetworkServices.PortalBackend.PortalEntities;
 using CatenaX.NetworkServices.PortalBackend.PortalEntities.Entities;
 using CatenaX.NetworkServices.Tests.Shared;
+using CatenaX.NetworkServices.Tests.Shared.Extensions;
 using FakeItEasy;
 using FluentAssertions;
 using Xunit;
