+++ tests/portalbackend/CatenaX.NetworkServices.PortalBackend.DBAccess.Tests/CompanyRespotitoryTests.cs
@@ -26,6 +26,7 @@ using CatenaX.NetworkServices.PortalBackend.PortalEntities;
 using CatenaX.NetworkServices.PortalBackend.PortalEntities.Entities;
 using CatenaX.NetworkServices.PortalBackend.PortalEntities.Enums;
 using CatenaX.NetworkServices.Tests.Shared;
+using CatenaX.NetworkServices.Tests.Shared.Extensions;
 using FakeItEasy;
 using FakeItEasy.Sdk;
 using FluentAssertions;
