+++ tests/portalbackend/CatenaX.NetworkServices.PortalBackend.DBAccess.Tests/UserRepositoryTests.cs
@@ -24,6 +24,7 @@ using CatenaX.NetworkServices.PortalBackend.DBAccess.Repositories;
 using CatenaX.NetworkServices.PortalBackend.PortalEntities;
 using CatenaX.NetworkServices.PortalBackend.PortalEntities.Entities;
 using CatenaX.NetworkServices.Tests.Shared;
+using CatenaX.NetworkServices.Tests.Shared.Extensions;
 using FakeItEasy;
 using FluentAssertions;
 using Xunit;
