+++ tests/shared/CatenaX.NetworkServices.Tests.Shared/IntegrationTests/FakePolicyEvaluator.cs
@@ -1,4 +1,24 @@
-﻿using System.Security.Claims;
+﻿/********************************************************************************
+ * Copyright (c) 2021,2022 BMW Group AG
+ * Copyright (c) 2021,2022 Contributors to the CatenaX (ng) GitHub Organisation.
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License, Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ********************************************************************************/
+
+using System.Security.Claims;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Authorization.Policy;
@@ -8,7 +28,7 @@ namespace CatenaX.NetworkServices.Tests.Shared.IntegrationTests;
 
 public class FakePolicyEvaluator : IPolicyEvaluator
 {
-    public virtual async Task<AuthenticateResult> AuthenticateAsync(AuthorizationPolicy policy, HttpContext context)
+    public async Task<AuthenticateResult> AuthenticateAsync(AuthorizationPolicy policy, HttpContext context)
     {
         var testScheme = "FakeScheme";
         var principal = new ClaimsPrincipal();
@@ -22,8 +42,8 @@ public class FakePolicyEvaluator : IPolicyEvaluator
             new AuthenticationProperties(), testScheme)));
     }
 
-    public virtual async Task<PolicyAuthorizationResult> AuthorizeAsync(AuthorizationPolicy policy,
-        AuthenticateResult authenticationResult, HttpContext context, object resource)
+    public async Task<PolicyAuthorizationResult> AuthorizeAsync(AuthorizationPolicy policy,
+        AuthenticateResult authenticationResult, HttpContext context, object? resource)
     {
         return await Task.FromResult(PolicyAuthorizationResult.Success());
     }
